working on currency data for invoice

+++ db/migrate/20210506153214_create_invoice_line_items.rb
@@ -2,7 +2,7 @@ class CreateInvoiceLineItems < ActiveRecord::Migration[6.1]
   def change
     create_table :invoice_line_items do |t|
       t.text :description
-      t.integer :price
+      t.decimal :price
       t.references :invoice, null: false, foreign_key: true
       t.timestamps
     end
